refactor: move modules to module infrastruture
